trello/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rabbitmq==0.2.0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
